compose: Splitting trying one by one.
