+++ static/js/compose.js
@@ -14,9 +14,9 @@ import * as common from "./common";
 import * as compose_actions from "./compose_actions";
 import * as compose_error from "./compose_error";
 import * as compose_fade from "./compose_fade";
-import * as compose_pm_pill from "./compose_pm_pill";
 import * as compose_state from "./compose_state";
 import * as compose_ui from "./compose_ui";
+import * as compose_validate from "./compose_validate";
 import * as drafts from "./drafts";
 import * as echo from "./echo";
 import * as giphy from "./giphy";
@@ -58,9 +58,9 @@ import * as zcommand from "./zcommand";
 
 let user_acknowledged_all_everyone;
 let user_acknowledged_announce;
-let wildcard_mention;
 let uppy;
 
+export let wildcard_mention;
 export let wildcard_mention_large_stream_threshold = 15;
 export const announce_warn_threshold = 60;
 export const uploads_domain = document.location.protocol + "//" + document.location.host;
@@ -70,6 +70,10 @@ export const uploads_re = new RegExp(
     "\\]\\(" + uploads_domain + "(" + uploads_path + "[^\\)]+)\\)",
     "g",
 );
+
+export function set_wildcard_mention(value) {
+    wildcard_mention = value;
+}
 
 function make_uploads_relative(content) {
     // Rewrite uploads in Markdown links back to domain-relative form
@@ -122,7 +126,7 @@ export function clear_all_everyone_warnings() {
     $("#compose-send-status").hide();
 }
 
-function show_sending_indicator(whats_happening) {
+export function show_sending_indicator(whats_happening) {
     $("#sending-indicator").text(whats_happening);
     $("#sending-indicator").show();
 }
@@ -363,7 +367,7 @@ export function finish() {
         return undefined;
     }
 
-    if (!validate()) {
+    if (!compose_validate.validate()) {
         return false;
     }
 
@@ -493,7 +497,7 @@ export function set_wildcard_mention_large_stream_threshold(value) {
     wildcard_mention_large_stream_threshold = value;
 }
 
-function validate_stream_message_mentions(stream_id) {
+export function validate_stream_message_mentions(stream_id) {
     const stream_count = peer_data.get_subscriber_count(stream_id) || 0;
 
     // If the user is attempting to do a wildcard mention in a large
@@ -530,7 +534,7 @@ function validate_stream_message_mentions(stream_id) {
     return true;
 }
 
-function validate_stream_message_announce(sub) {
+export function validate_stream_message_announce(sub) {
     const stream_count = peer_data.get_subscriber_count(sub.stream_id) || 0;
 
     if (sub.name === "announce" && stream_count > announce_warn_threshold) {
@@ -551,7 +555,7 @@ function validate_stream_message_announce(sub) {
     return true;
 }
 
-function validate_stream_message_post_policy(sub) {
+export function validate_stream_message_post_policy(sub) {
     if (page_params.is_admin) {
         return true;
     }
@@ -653,145 +657,6 @@ export function validate_stream_message_address_info(stream_name) {
     const autosubscribe = page_params.narrow_stream !== undefined;
     const error_type = check_unsubscribed_stream_for_send(stream_name, autosubscribe);
     return validation_error(error_type, stream_name);
-}
-
-function validate_stream_message() {
-    const stream_name = compose_state.stream_name();
-    if (stream_name === "") {
-        compose_error.show(
-            $t_html({defaultMessage: "Please specify a stream"}),
-            $("#stream_message_recipient_stream"),
-        );
-        return false;
-    }
-
-    if (page_params.realm_mandatory_topics) {
-        const topic = compose_state.topic();
-        if (topic === "") {
-            compose_error.show(
-                $t_html({defaultMessage: "Please specify a topic"}),
-                $("#stream_message_recipient_topic"),
-            );
-            return false;
-        }
-    }
-
-    const sub = stream_data.get_sub(stream_name);
-    if (!sub) {
-        return validation_error("does-not-exist", stream_name);
-    }
-
-    if (!validate_stream_message_post_policy(sub)) {
-        return false;
-    }
-
-    /* Note: This is a global and thus accessible in the functions
-       below; it's important that we update this state here before
-       proceeding with further validation. */
-    wildcard_mention = util.find_wildcard_mentions(compose_state.message_content());
-
-    // If both `@all` is mentioned and it's in `#announce`, just validate
-    // for `@all`. Users shouldn't have to hit "yes" more than once.
-    if (wildcard_mention !== null && stream_name === "announce") {
-        if (
-            !validate_stream_message_address_info(stream_name) ||
-            !validate_stream_message_mentions(sub.stream_id)
-        ) {
-            return false;
-        }
-        // If either criteria isn't met, just do the normal validation.
-    } else {
-        if (
-            !validate_stream_message_address_info(stream_name) ||
-            !validate_stream_message_mentions(sub.stream_id) ||
-            !validate_stream_message_announce(sub)
-        ) {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-// The function checks whether the recipients are users of the realm or cross realm users (bots
-// for now)
-function validate_private_message() {
-    if (page_params.realm_private_message_policy === 2) {
-        // Frontend check for for PRIVATE_MESSAGE_POLICY_DISABLED
-        const user_ids = compose_pm_pill.get_user_ids();
-        if (user_ids.length !== 1 || !people.get_by_user_id(user_ids[0]).is_bot) {
-            // Unless we're composing to a bot
-            compose_error.show(
-                $t_html({defaultMessage: "Private messages are disabled in this organization."}),
-                $("#private_message_recipient"),
-            );
-            return false;
-        }
-    }
-
-    if (compose_state.private_message_recipient().length === 0) {
-        compose_error.show(
-            $t_html({defaultMessage: "Please specify at least one valid recipient"}),
-            $("#private_message_recipient"),
-        );
-        return false;
-    } else if (page_params.realm_is_zephyr_mirror_realm) {
-        // For Zephyr mirroring realms, the frontend doesn't know which users exist
-        return true;
-    }
-
-    const invalid_recipients = get_invalid_recipient_emails();
-
-    let context = {};
-    if (invalid_recipients.length === 1) {
-        context = {recipient: invalid_recipients.join(",")};
-        compose_error.show(
-            $t_html({defaultMessage: "The recipient {recipient} is not valid"}, context),
-            $("#private_message_recipient"),
-        );
-        return false;
-    } else if (invalid_recipients.length > 1) {
-        context = {recipients: invalid_recipients.join(",")};
-        compose_error.show(
-            $t_html({defaultMessage: "The recipients {recipients} are not valid"}, context),
-            $("#private_message_recipient"),
-        );
-        return false;
-    }
-    return true;
-}
-
-export function validate() {
-    $("#compose-send-button").prop("disabled", true).trigger("blur");
-    const message_content = compose_state.message_content();
-    if (reminder.is_deferred_delivery(message_content)) {
-        show_sending_indicator($t({defaultMessage: "Scheduling..."}));
-    } else {
-        show_sending_indicator($t({defaultMessage: "Sending..."}));
-    }
-
-    if (/^\s*$/.test(message_content)) {
-        compose_error.show(
-            $t_html({defaultMessage: "You have nothing to send!"}),
-            $("#compose-textarea"),
-        );
-        return false;
-    }
-
-    if ($("#zephyr-mirror-error").is(":visible")) {
-        compose_error.show(
-            $t_html({
-                defaultMessage:
-                    "You need to be running Zephyr mirroring in order to send messages!",
-            }),
-        );
-        return false;
-    }
-
-    if (compose_state.get_message_type() === "private") {
-        return validate_private_message();
-    }
-    return validate_stream_message();
 }
 
 export function handle_keydown(event, textarea) {
